Produce relevant tags for coding problem solutions
This initial commit contains code that interprets the solution code
for common coding problems (such as those found on LeetCode) contained
in local markdown files and produces relevant tags for each solution.
Tags are appended to the frontmatter of the selected markdown file.
Edge cases are properly handled to account for the presence or absence
of pre-existing tags in the markdown file.